# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: YouTillEyes Marketplace Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin only)
      parameters:
        - in: query
          name: role
          schema:
            type: string
            enum: [ADMIN, CLIENT, USER]
        - in: query
          name: search
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get user by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /submissions:
    get:
      operationId: getSubmissions
      tags: [submissions]
      summary: Get all submissions (admin only)
      parameters:
        - in: query
          name: projectId
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [Pending, Approved, Forwarded, Rejected]
      responses:
        "200":
          description: List of submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
    post:
      operationId: createSubmission
      tags: [submissions]
      summary: Submit work for a project (user only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSubmissionRequest"
      responses:
        "201":
          description: Submission created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
  /submissions/{id}:
    get:
      operationId: getSubmissionById
      tags: [submissions]
      summary: Get submission by ID
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Submission details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
    put:
      operationId: updateSubmission
      tags: [submissions]
      summary: Update submission status (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSubmissionRequest"
      responses:
        "200":
          description: Submission updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Submission"
...
  /submissions/project/{projectId}:
    get:
      operationId: getSubmissionsByProject
      tags: [submissions]
      summary: Get all submissions for a project
      parameters:
        - in: path
          name: projectId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project submissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Submission"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, CLIENT, USER]
        phone:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        skills:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - role
        - isActive
        - createdAt
    Project:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        category:
          type: string
          nullable: true
        budget:
          type: number
          nullable: true
        deadline:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [Open, Assigned, In-Progress, Submitted, Completed]
        clientId:
          type: integer
        assignedUserId:
          type: integer
          nullable: true
        client:
          $ref: "#/components/schemas/User"
          nullable: true
        assignedUser:
          $ref: "#/components/schemas/User"
          nullable: true
        bidCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - status
        - clientId
        - bidCount
        - createdAt
        - updatedAt
...
    Submission:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        userId:
          type: integer
        fileUrl:
          type: string
        fileDescription:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        status:
          type: string
          enum: [Pending, Approved, Forwarded, Rejected]
        adminNotes:
          type: string
          nullable: true
        project:
          $ref: "#/components/schemas/Project"
          nullable: true
        user:
          $ref: "#/components/schemas/User"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - userId
        - fileUrl
        - status
        - createdAt
        - updatedAt
    CreateSubmissionRequest:
      type: object
      properties:
        projectId:
          type: integer
        fileUrl:
          type: string
        fileDescription:
          type: string
        notes:
          type: string
      required:
        - projectId
        - fileUrl
    UpdateSubmissionRequest:
      type: object
      properties:
        status:
          type: string
          enum: [Pending, Approved, Forwarded, Rejected]
        adminNotes:
          type: string
...