# lib/api-zod/src/generated/types/createSubmissionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * YouTillEyes Marketplace Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateSubmissionRequest {
  projectId: number;
  fileUrl: string;
  fileDescription?: string;
  notes?: string;
}